feat: complete modules tasks
